+++ aksara/__init__.py
@@ -44,6 +44,15 @@ from aksara.i18n import (
     reset_timezone,
     translate,
 )
+from aksara.contenttypes import (
+    ContentType,
+    clear_content_type_cache,
+    get_content_type_by_id,
+    get_content_type_for_model,
+    get_model_for_content_type,
+    sync_content_types,
+)
+from aksara.workflows import DurableStep, DurableStepState
 from aksara.storage import FieldFile, Storage, FileSystemStorage, S3Storage, get_default_storage
 from aksara.exceptions import (
     AksaraError,
@@ -244,6 +253,14 @@ __all__ = [
     "reset_outbox",
     "reset_timezone",
     "translate",
+    "ContentType",
+    "clear_content_type_cache",
+    "get_content_type_by_id",
+    "get_content_type_for_model",
+    "get_model_for_content_type",
+    "sync_content_types",
+    "DurableStep",
+    "DurableStepState",
     # Exceptions
     "AksaraError",
     "DatabaseError",